helper update fixed

+++ files/helper.py
@@ -2,14 +2,15 @@ import os
 from .models import File,Image
 from django.conf import settings
 
-def classification_helper(files,form,lawyer):       
+def classification_helper(files,form,lawyer,update):       
         file_1 = form.save(commit=False)
         file_1.lawyer=lawyer
         file_1.save()
         file_name = File.objects.filter(id=file_1.id).first()   
         for file in files:
-            if (len(file.name) > 180):               
-                file_1.delete()               
+            if (len(file.name) > 180):
+                if (update != True):                
+                    file_1.delete()               
                 print(len(file.name))
                 return False
             deneme= Image.objects.create(file_name=file_name, image=file)
@@ -26,4 +27,3 @@ def classification_helper(files,form,lawyer):
             os.replace(initial_path, new_path)
             deneme.image = link_path 
             deneme.save()
-
